requirements.txt
kivy
bleak
pyjnius
